update
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@367 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ kernel/fiasco/src/kern/setchecksum.c
@@ -7,7 +7,7 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <string.h>
-
+#include <inttypes.h>
 
 static void usage(void)
 {
@@ -24,9 +24,9 @@ int main(int argc, char *argv[])
 {
 
   void *base;
-  unsigned long volatile *chksum;
+  uint32_t volatile *chksum;
   unsigned long left;
-  
+
   if (argc!=4)
     {
       usage();
